+++ src/postgres/discovery/executor/mock.rs
@@ -1,7 +1,7 @@
 use crate::sqlx_types::{postgres::PgRow, PgPool};
 use sea_query::{PostgresQueryBuilder, SelectStatement};
 
-use crate::debug_print;
+use crate::{debug_print, sqlx_types::SqlxError};
 
 #[allow(dead_code)]
 pub struct Executor {
@@ -19,7 +19,7 @@ impl IntoExecutor for PgPool {
 }
 
 impl Executor {
-    pub async fn fetch_all(&self, select: SelectStatement) -> Vec<PgRow> {
+    pub async fn fetch_all(&self, select: SelectStatement) -> Result<Vec<PgRow>, SqlxError> {
         let (_sql, _values) = select.build(PostgresQueryBuilder);
         debug_print!("{}, {:?}", _sql, _values);